+++ frontend/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.27.0](https://github.com/bcit-tlu/hriv/compare/frontend-v0.26.0...frontend-v0.27.0) (2026-06-24)
+
+
+### Features
+
+* **backend:** track tile-cache provenance and staleness ([#747](https://github.com/bcit-tlu/hriv/issues/747)) ([44d2c23](https://github.com/bcit-tlu/hriv/commit/44d2c2354ef4020744c39ee19e5e72ab48eecb4d))
+
+
+### Bug Fixes
+
+* move category restriction reset ([#741](https://github.com/bcit-tlu/hriv/issues/741)) ([b7f8e9a](https://github.com/bcit-tlu/hriv/commit/b7f8e9a2719c3c598bb68ebe0ad20a16b8eb4ce9))
+
+## [0.26.0](https://github.com/bcit-tlu/hriv/compare/frontend-v0.25.0...frontend-v0.26.0) (2026-06-22)
+
+
+### Features
+
+* **frontend:** share category picker collapse state ([#729](https://github.com/bcit-tlu/hriv/issues/729)) ([ced2883](https://github.com/bcit-tlu/hriv/commit/ced28839b6c4978dbda9dfa4150aeb2d0a00468c))
+
 ## [0.25.0](https://github.com/bcit-tlu/hriv/compare/frontend-v0.24.0...frontend-v0.25.0) (2026-06-19)